bot_tg/requirements.txt
aiogram
requests
aiohttp
psycopg2-binary